Add support for DOT, UMLGraph and PlantUML
This commit will enable a user to use DOT, UML and PlantUML
diagrams in his website.

Diagrams are rendered as SVG images by a service called Gravizo on their
servers. It supports DOT, UML and PlantUML syntax. It also has it own
JSON based diagram syntax.

I chose to use gravizo as opposed to Mermaid.js or Chart.js for the
following reasons:

- It's a Server Side service, and hence no JS.
- It uses familiar and well used Syntax.
- It uses AT&T's graphviz as a backend.