Add support for x/y graphs
This commit will allow a user to represent x/y type graphs in his
website.

A user uses the `chartjs` shorcode for charts. The charts are written in
chartjs syntax and then rendered as PNG, irrespective of whether a
viewer supports Javascript or not.

2 commits back, I added support for DOT, UMLGraph, and PlantUML syntax.
The problem with theses sytaxes is that they are meant for node/edge
graphs. A user of the theme will not be able to represent data in the
form of a Pie Chart, Bar Chart, or Line Chart using the afore mentioned
syntaxes.

Hence, I have decided to use Chartjs syntax for this. I however, have
chosen **not** to use Chartjs for rendering the charts. I have instead,
opted to use a service called QuickChart for rendering the charts.

QuickChart is a server-side renderer that supports the Chartjs syntax
and also, the deprecated Google Image Charts syntax as well. I have set
things such that it returns a PNG. This PNG is then inserted to the
respective page.

Useful Links:

- [Chartjs](www.chartjs.org)
- [QuickChart](www.quickchart.io)
- [Node to Edge graphs](https://en.wikipedia.org/wiki/Graph_theory)
- [X/Y charts](https://en.wikipedia.org/wiki/Chart)