Add more configurable options for math
This commit will enable a user to:

- Use only images to render math (disable katex)
- Use katex on all pages
- Use katex on a per-page basis

I personally feel it's a must to have a little bit of configurability in
our software so that Users can configure the software to their taste.
Though please not that this option is not for the sake of
configurability.

After this commit, one will access all math rendering options at
`markup.math` (with the exception of the per page rendering). As of now,
the following options are used:

- markup.math.katex (Boolean, Whether to use katex or not)
- markup.math.allPages (Boolean, Whether to use katex on every page)
- .Params.math (Boolean, Enable katex for a specific page)