TeX Shortcode: Use .Inner to get equation
This commit will enable a user to have multiline shortcodes.

Previously, equations were passed as strings to Hugo. The problem with
is that Hugo doesn't support multiline strings.

The reason I needed strings in the first place, was because I needed to
assign the string to the parameter `display`, as a sort of flag. I would
then check if `display` existed, and do things accordingly.

Now, I pass a Boolean to `display` if I need display mode, and pass the
equation as `.Inner` (directly in the shortcode without the use of a
string)

I however, only try to get the value of `display` if it exists. This
means I do not need to pass `display=false` every time I need to use
inline mode.